Move tests out of production code
Into a separate package.
