+++ setup.cfg
@@ -2,6 +2,6 @@
 license-file = LICENSE.txt
 
 [tool:pytest]
-DJANGO_SETTINGS_MODULE = rest_framework_json_schema.dummy.settings
+DJANGO_SETTINGS_MODULE = tests.dummy.settings
 flake8-max-line-length = 100
 flake8-max-complexity = 15
